Batch and cell update token amount and chain
