+++ lib/focalboard/card.ts
@@ -4,6 +4,7 @@ import type { Block } from 'lib/focalboard/block';
 import { createBlock } from 'lib/focalboard/block';
 import type { ProposalWithUsersLite } from 'lib/proposal/getProposals';
 import type { ProposalEvaluationStatus } from 'lib/proposal/interface';
+import type { RewardType } from 'lib/rewards/interfaces';
 
 export type CardPropertyValue = string | string[] | number;
 
@@ -32,6 +33,7 @@ export type CardPageProposal = {
 
 export type CardPageReward = {
   id: string;
+  rewardType: RewardType;
 };
 
 export type CardPage<V = CardPropertyValue> = {
